+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 
 ## 4.3.0 (roadmap)
 
-- [x] `createExtension` accepts referenced Target names as in `$TARGET_NAME`
+- [x] [core] `createExtension` accepts referenced Target names as in `$TARGET_NAME`
+- [x] [core] deprecate `getAction` for `getTarget`
 - [ ] Should be able to `exports.name = 'featureName'; exports.actions = [{ hook, handler }, ...]`
 - [ ] Add a fullstack test in `service-fetchq` with a task entered by a route and then a polling to monitor completion of such task
 - [ ] Fix Apollo Client service