+++ packages/core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v4.3.0
 
 - `createExtension` accepts referenced Target names as in `$TARGET_NAME`
+- deprecate `getAction` for `getTarget`
 
 ## v4.2.0